add a local worker policy
