+++ dora/core/client/fs/src/main/java/alluxio/client/block/BlockWorkerInfo.java
@@ -24,6 +24,7 @@ import javax.annotation.concurrent.ThreadSafe;
  */
 @PublicApi
 @ThreadSafe
+// TODO(jiacheng): this class should just be renamed!
 public final class BlockWorkerInfo {
   private final WorkerNetAddress mNetAddress;
   private final long mCapacityBytes;
